Isolate Mission with new interface
Convert Misison to an interface to clearly defintion the interactino
with the Simulation. The existing Misison class is an asbtract class.
This will allow the Misison to have different implementation without
affecting the simulation.

Changes:
- Mission class changed to AbstractMission class
- New Mission interface defines interaction with callers.
- Mission description is prealced with name as the 2 are the same value.
- Create a MisisonLog class to encapsulate everything historical about a
  Mission
- Key dates for a Mission are contained within the MisisonLog
- Reduce the overlapping methods to ensure Mission interface is the core
  set of memthods.

Prework for #625
