+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/person/ai/mission/meta/RescueSalvageVehicleMeta.java
@@ -85,7 +85,7 @@ public class RescueSalvageVehicleMeta extends AbstractMetaMission {
             // FIXME : need to know how many extra EVA suits needed in the broken vehicle
 
             // Check if min number of EVA suits at settlement.
-            if (Mission.getNumberAvailableEVASuitsAtSettlement(settlement) < min_num) {
+            if (RoverMission.getNumberAvailableEVASuitsAtSettlement(settlement) < min_num) {
     	        return 0;
     	    }
