+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/person/ai/task/ReviewMissionPlan.java
@@ -216,7 +216,7 @@ public class ReviewMissionPlan extends Task implements Serializable {
 						    if (is90Completed()) {
 						    	
 								logger.log(worker, Level.INFO, 15_000, "Reviewing " + requestedBy
-										+ "'s " + m.getDescription() + " mission plan.");
+										+ "'s " + m.getName() + " mission plan.");
 								
 				            	// Use up to 90% of the time
 								return 0; 
@@ -399,7 +399,7 @@ public class ReviewMissionPlan extends Task implements Serializable {
 								missionManager.scoreMissionPlan(mp, score, person);
 
 								StringBuilder msg = new StringBuilder();
-								msg.append("Grading ").append(requestedBy).append("'s ").append(m.getDescription());
+								msg.append("Grading ").append(requestedBy).append("'s ").append(m.getName());
 								msg.append(" plan - ");
 								msg.append(" Rating: ").append(rating); 
 								msg.append(", Rels: ").append(relation); 
@@ -510,7 +510,7 @@ public class ReviewMissionPlan extends Task implements Serializable {
 							missionManager.approveMissionPlan(mp, p, PlanType.APPROVED, settlement.getMinimumPassingScore());
 								
 							logger.log(worker, Level.INFO, 0, "Approved " + requestedBy
-									+ "'s " + m.getDescription() + " mission plan. Total Score: " 
+									+ "'s " + m.getName() + " mission plan. Total Score: " 
 									+ Math.round(score*10.0)/10.0 
 									+ " [Min: " + Math.round(settlement.getMinimumPassingScore()*10.0)/10.0 + "].");
 						} else {
@@ -519,7 +519,7 @@ public class ReviewMissionPlan extends Task implements Serializable {
 							missionManager.approveMissionPlan(mp, p, PlanType.NOT_APPROVED, settlement.getMinimumPassingScore());
 						
 							logger.log(worker, Level.INFO, 0, "Did NOT approve " + requestedBy
-									+ "'s " + m.getDescription() + " mission plan. Total Score: " 
+									+ "'s " + m.getName() + " mission plan. Total Score: " 
 									+ Math.round(score*10.0)/10.0 
 									+ " [Min: " + Math.round(settlement.getMinimumPassingScore()*10.0)/10.0 + "].");
 						}
