+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/person/health/DeathInfo.java
@@ -215,7 +215,7 @@ public class DeathInfo implements Serializable {
 
 		Mission mm = mind.getMission();
 		if (mm != null) {
-			mission = mm.getDescription();
+			mission = mm.getName();
 			missionPhase = mm.getPhaseDescription();
 		}
