+++ mars-sim-ui/src/main/java/org/mars_sim/msp/ui/swing/tool/mission/create/ConstructionVehiclePanel.java
@@ -245,7 +245,7 @@ class ConstructionVehiclePanel extends WizardPanel {
                         result = vehicle.printStatusTypes();
                     else if (column == 2) {
                         Mission mission = missionManager.getMissionForVehicle(vehicle);
-                        if (mission != null) result = mission.getDescription();
+                        if (mission != null) result = mission.getName();
                         else result = "None";
                     }
                 }
