+++ mars-sim-ui/src/main/java/org/mars_sim/msp/ui/swing/tool/mission/create/VehiclePanel.java
@@ -255,7 +255,7 @@ class VehiclePanel extends WizardPanel {
 					else if (column == 9) {
 						Mission mission = missionManager.getMissionForVehicle(vehicle);
 						if (mission != null)
-							result = mission.getDescription();
+							result = mission.getName();
 						else
 							result = "None";
 					}
